aider: Add the ability to run any custom command.
# Context:
USER: Manage the power status and service status of windows servers, remotely.  Use powershell commands to save time.
ASSISTANT: To manage the power status and service status of Windows servers remotely using PowerShell commands, you'll need to use the `Invoke-Command` cmdlet, which runs commands on a local or remote computer.

Here's a basic example of how you can use PowerShell to manage a remote server:

```powershell
# Define the server name
$serverName = "Server01"

# Define the command to run
$command = {
    # Get the status of a service
    Get-Service -Name "wuauserv"
}

# Run the command on the remote server
Invoke-Command -ComputerName $serverName -ScriptBlock $command
```

This script gets the status of the Windows Update service (`wuauserv`) on a remote server (`Server01`).

You can replace `"wuauserv"` with the name of the service you want to manage, and `"Server01"` with the name of your server.

To manage the power status of the server, you can use the `Stop-Computer` cmdlet to shut down the server, or the `Restart-Computer` cmdlet to restart it. Here's an example:

```powershell
# Define the server name
$serverName = "Server01"

# Shut down the server
Invoke-Command -ComputerName $serverName -ScriptBlock { Stop-Computer -Force }

# Restart the server
Invoke-Command -ComputerName $serverName -ScriptBlock { Restart-Computer -Force }
```

Again, replace `"Server01"` with the name of your server.

Please note that to run commands on a remote server, you'll need to have the appropriate permissions on the server, and the server will need to be configured to allow remote PowerShell sessions.
USER: Add the ability to run any custom command
ASSISTANT: To add the ability to run any custom command, you can modify the script to accept the command as a parameter. Here's how you can do it:

```python
app.py
<<<<<<< ORIGINAL

=======
# Import the required module
import subprocess

# Define the server name
serverName = "Server01"

# Define the command to run
command = input("Enter the command to run: ")

# Run the command on the remote server
subprocess.run(["powershell", "-Command", f"Invoke-Command -ComputerName {serverName} -ScriptBlock {{ {command} }}"])
>>>>>>> UPDATED
```

In this updated script, the `input` function is used to get the command to run from the user. The `subprocess.run` function is then used to run the command on the remote server.

Please note that this script will run the command on the remote server as is, without any validation or error checking. Depending on the command, this could potentially be dangerous. You should add appropriate error checking and validation to ensure that only valid and safe commands are run.